+++ src/validators/required.ts
@@ -1,8 +1,8 @@
 import { ValidatorFn } from './types'
 
-const requiredValidator: ValidatorFn = () => ({
+const requiredValidator: ValidatorFn = required => ({
   error: 'REQUIRED',
-  validate: (value: unknown) => !!value,
+  validate: (value: unknown) => (required ? !!value : true),
 })
 
 export default requiredValidator